Update README.md - repo note
## repository
The original repository MCBCMF/MCBCMassacre was setup in a way that it cannot be deleted and updates can occur from only one secured device. We consider it the vault. 

Due to recent activity, we've suspended updates to our fork and are only making commits there. 

If you haven't already, please clone the repository and fetch/pull often. The more clones we have the more difficult this becomes to keep hidden.